requirements.txt
colorlog>=6.8.0
beautifulsoup4>=4.12.0
reactionmenu>=3.1.7
pillow>=10.4.0
requests>=2.32.0
cexprtk>=0.4.0
discord.py>=2.4.0
rollplayerlib==0.5.6
peridata>=1.3.3
